Python/NyckelRegister/requirements.txt
tkinter
sqlite3
odfpy
